Refactor demo/test types
Collapsed three virtually identical `NodeType` implementations into one,
split code from `testing::demo` into more descriptive modules and
renamed various test types to conform to a single naming scheme.
