+++ crates/hotshot/examples/combined/types.rs
@@ -1,6 +1,6 @@
 use crate::infra::CombinedDARun;
 use hotshot::traits::implementations::{CombinedCommChannel, MemoryStorage};
-use hotshot_testing::demo::DemoTypes;
+use hotshot_testing::state_types::TestTypes;
 use hotshot_types::traits::node_implementation::{ChannelMaps, NodeImplementation, NodeType};
 use serde::{Deserialize, Serialize};
 use std::fmt::Debug;
@@ -8,20 +8,20 @@ use std::fmt::Debug;
 #[derive(Clone, Debug, Deserialize, Serialize, Hash, PartialEq, Eq)]
 pub struct NodeImpl {}
 
-pub type DANetwork = CombinedCommChannel<DemoTypes>;
-pub type VIDNetwork = CombinedCommChannel<DemoTypes>;
-pub type QuorumNetwork = CombinedCommChannel<DemoTypes>;
-pub type ViewSyncNetwork = CombinedCommChannel<DemoTypes>;
+pub type DANetwork = CombinedCommChannel<TestTypes>;
+pub type VIDNetwork = CombinedCommChannel<TestTypes>;
+pub type QuorumNetwork = CombinedCommChannel<TestTypes>;
+pub type ViewSyncNetwork = CombinedCommChannel<TestTypes>;
 
-impl NodeImplementation<DemoTypes> for NodeImpl {
-    type Storage = MemoryStorage<DemoTypes>;
+impl NodeImplementation<TestTypes> for NodeImpl {
+    type Storage = MemoryStorage<TestTypes>;
     type QuorumNetwork = QuorumNetwork;
     type CommitteeNetwork = DANetwork;
 
     fn new_channel_maps(
-        start_view: <DemoTypes as NodeType>::Time,
-    ) -> (ChannelMaps<DemoTypes>, Option<ChannelMaps<DemoTypes>>) {
+        start_view: <TestTypes as NodeType>::Time,
+    ) -> (ChannelMaps<TestTypes>, Option<ChannelMaps<TestTypes>>) {
         (ChannelMaps::new(start_view), None)
     }
 }
-pub type ThisRun = CombinedDARun<DemoTypes>;
+pub type ThisRun = CombinedDARun<TestTypes>;
