+++ crates/hotshot/examples/infra/mod.rs
@@ -22,7 +22,7 @@ use hotshot_orchestrator::{
     config::{NetworkConfig, NetworkConfigFile, WebServerConfig},
 };
 use hotshot_task::task::FilterEvent;
-use hotshot_testing::demo::block::{VIDBlockHeader, VIDBlockPayload, VIDTransaction};
+use hotshot_testing::block_types::{TestBlockHeader, TestBlockPayload, TestTransaction};
 use hotshot_types::message::Message;
 use hotshot_types::traits::network::ConnectedNetwork;
 use hotshot_types::ValidatorConfig;
@@ -296,7 +296,7 @@ pub trait RunDA<
 > where
     <TYPES as NodeType>::StateType: TestableState,
     <TYPES as NodeType>::BlockPayload: TestableBlock,
-    TYPES: NodeType<Transaction = VIDTransaction>,
+    TYPES: NodeType<Transaction = TestTransaction>,
     Leaf<TYPES>: TestableLeaf,
     Self: Sync,
     SystemContext<TYPES, NODE>: HotShotType<TYPES, NODE>,
@@ -377,7 +377,7 @@ pub trait RunDA<
     async fn run_hotshot(
         &self,
         mut context: SystemContextHandle<TYPES, NODE>,
-        transactions: &mut Vec<VIDTransaction>,
+        transactions: &mut Vec<TestTransaction>,
         transactions_to_send_per_round: u64,
     ) {
         let NetworkConfig {
@@ -498,9 +498,9 @@ pub struct WebServerDARun<TYPES: NodeType> {
 #[async_trait]
 impl<
         TYPES: NodeType<
-            Transaction = VIDTransaction,
-            BlockPayload = VIDBlockPayload,
-            BlockHeader = VIDBlockHeader,
+            Transaction = TestTransaction,
+            BlockPayload = TestBlockPayload,
+            BlockHeader = TestBlockHeader,
         >,
         NODE: NodeImplementation<
             TYPES,
@@ -601,9 +601,9 @@ pub struct Libp2pDARun<TYPES: NodeType> {
 #[async_trait]
 impl<
         TYPES: NodeType<
-            Transaction = VIDTransaction,
-            BlockPayload = VIDBlockPayload,
-            BlockHeader = VIDBlockHeader,
+            Transaction = TestTransaction,
+            BlockPayload = TestBlockPayload,
+            BlockHeader = TestBlockHeader,
         >,
         NODE: NodeImplementation<
             TYPES,
@@ -694,9 +694,9 @@ pub struct CombinedDARun<TYPES: NodeType> {
 #[async_trait]
 impl<
         TYPES: NodeType<
-            Transaction = VIDTransaction,
-            BlockPayload = VIDBlockPayload,
-            BlockHeader = VIDBlockHeader,
+            Transaction = TestTransaction,
+            BlockPayload = TestBlockPayload,
+            BlockHeader = TestBlockHeader,
         >,
         NODE: NodeImplementation<
             TYPES,
@@ -806,9 +806,9 @@ where
 /// Main entry point for validators
 pub async fn main_entry_point<
     TYPES: NodeType<
-        Transaction = VIDTransaction,
-        BlockPayload = VIDBlockPayload,
-        BlockHeader = VIDBlockHeader,
+        Transaction = TestTransaction,
+        BlockPayload = TestBlockPayload,
+        BlockHeader = TestBlockHeader,
     >,
     DACHANNEL: CommunicationChannel<TYPES> + Debug,
     QUORUMCHANNEL: CommunicationChannel<TYPES> + Debug,
