+++ crates/hotshot/examples/libp2p/validator.rs
@@ -1,6 +1,6 @@
 use async_compatibility_layer::logging::{setup_backtrace, setup_logging};
 use clap::Parser;
-use hotshot_testing::demo::DemoTypes;
+use hotshot_testing::state_types::TestTypes;
 use tracing::{info, instrument};
 use types::VIDNetwork;
 
@@ -28,7 +28,7 @@ async fn main() {
         args.url, args.port
     );
     infra::main_entry_point::<
-        DemoTypes,
+        TestTypes,
         DANetwork,
         QuorumNetwork,
         ViewSyncNetwork,
