+++ crates/testing/src/node_types.rs
@@ -1,8 +1,8 @@
 use hotshot::traits::election::static_committee::GeneralStaticCommittee;
 
-use crate::demo::{
-    block::{VIDBlockHeader, VIDBlockPayload, VIDTransaction},
-    DemoState,
+use crate::{
+    block_types::{TestBlockHeader, TestBlockPayload, TestTransaction},
+    state_types::TestState,
 };
 
 use hotshot::{
@@ -38,12 +38,12 @@ use serde::{Deserialize, Serialize};
 pub struct TestTypes;
 impl NodeType for TestTypes {
     type Time = ViewNumber;
-    type BlockHeader = VIDBlockHeader;
-    type BlockPayload = VIDBlockPayload;
+    type BlockHeader = TestBlockHeader;
+    type BlockPayload = TestBlockPayload;
     type SignatureKey = BLSPubKey;
-    type Transaction = VIDTransaction;
+    type Transaction = TestTransaction;
     type ElectionConfigType = StaticElectionConfig;
-    type StateType = DemoState;
+    type StateType = TestState;
     type Membership = GeneralStaticCommittee<TestTypes, Self::SignatureKey>;
 }
