+++ crates/testing/src/task_helpers.rs
@@ -1,7 +1,7 @@
 use std::marker::PhantomData;
 
 use crate::{
-    demo::block::{VIDBlockHeader, VIDBlockPayload},
+    block_types::{TestBlockHeader, TestBlockPayload},
     node_types::{MemoryImpl, TestTypes},
     test_builder::TestMetadata,
 };
@@ -125,9 +125,9 @@ async fn build_quorum_proposal_and_signature(
     let parent_leaf = leaf.clone();
 
     // every event input is seen on the event stream in the output.
-    let block = <VIDBlockPayload as TestableBlock>::genesis();
+    let block = <TestBlockPayload as TestableBlock>::genesis();
     let payload_commitment = vid_commitment(&block.encode().unwrap().collect());
-    let block_header = VIDBlockHeader::new(payload_commitment, (), &parent_leaf.block_header);
+    let block_header = TestBlockHeader::new(payload_commitment, (), &parent_leaf.block_header);
     let leaf = Leaf {
         view_number: ViewNumber::new(view),
         justify_qc: consensus.high_qc.clone(),
