+++ crates/testing/tests/network_task.rs
@@ -19,7 +19,7 @@ use std::{collections::HashMap, marker::PhantomData};
 #[ignore]
 async fn test_network_task() {
     use hotshot_task_impls::harness::run_harness;
-    use hotshot_testing::{demo::block::VIDTransaction, task_helpers::build_system_handle};
+    use hotshot_testing::{block_types::TestTransaction, task_helpers::build_system_handle};
     use hotshot_types::{data::VidDisperse, message::Proposal};
 
     async_compatibility_layer::logging::setup_logging();
@@ -33,8 +33,8 @@ async fn test_network_task() {
     let pub_key = *api.public_key();
     let priv_key = api.private_key();
     let vid = vid_init();
-    let transactions = vec![VIDTransaction(vec![0])];
-    let encoded_transactions = VIDTransaction::encode(transactions.clone()).unwrap();
+    let transactions = vec![TestTransaction(vec![0])];
+    let encoded_transactions = TestTransaction::encode(transactions.clone()).unwrap();
     let vid_disperse = vid.disperse(&encoded_transactions).unwrap();
     let payload_commitment = vid_disperse.commit;
     let signature =
